+++ client/src/Actions/login.js
@@ -3,7 +3,7 @@ const axios = require("axios");
 export async function login(password) {
   var response;
   await axios
-    .post("/api/login/", { password: password })
+    .post("http://localhost:5000/api/login/", { password: password })
     .then((res) => (response = res.data));
   return response;
 }
