+++ client/src/App.css
@@ -14,8 +14,11 @@
 }
 
 h1,
+h3 {
+  color: #6980e9;
+}
+
 h2,
-h3,
 h4,
 h5,
 h6,
