+++ client/src/css/Components/MessageList.css
@@ -1,0 +1,3 @@
+.delete:hover {
+  cursor: pointer;
+}
