+++ client/src/css/Components/Project.css
@@ -41,7 +41,6 @@
 
 .project-img:hover {
   transform: scale(1.05);
-  box-shadow: 5px 10px 18px #3b5998;
 }
 
 .vertical-center {
