+++ client/src/css/Pages/Admin.css
@@ -2,7 +2,7 @@
   text-align: center;
   background: #252934;
   width: 100%;
-  height: 50.5vw;
+  height: 100vw;
 }
 .header {
   color: white;
