+++ debug.log
@@ -35,3 +35,50 @@
 [0507/132738.415:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [0507/132738.415:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [0507/132738.415:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.306:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.307:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.307:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.308:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.308:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.308:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.308:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.308:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.308:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.308:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.308:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.308:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.308:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.308:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.308:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.309:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.309:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.309:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.309:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.309:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.309:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.309:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.309:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.309:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.309:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.309:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.309:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.309:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.309:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.309:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.309:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.310:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.310:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.310:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.310:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.310:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.310:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.310:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.310:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.310:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.310:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.310:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.311:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.311:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.311:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.311:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[0920/164811.311:ERROR:crash_report_database_win.cc(469)] failed to stat report
